Fix: alembic 버전 삭제 후 다시 생성
