--- watchapedia/database/alembic/versions/2025_01_30_0014-162001d72482_participant_like_count.py
@@ -1,38 +1,0 @@
-"""participant like_count
-
-Revision ID: 162001d72482
-Revises: 22af0add3b7c
-Create Date: 2025-01-30 00:14:49.272967
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = '162001d72482'
-down_revision: Union[str, None] = '22af0add3b7c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('participant', 'user_id')
-    op.drop_constraint('user_likes_participant_ibfk_2', 'user_likes_participant', type_='foreignkey')
-    op.drop_constraint('user_likes_participant_ibfk_1', 'user_likes_participant', type_='foreignkey')
-    op.create_foreign_key(None, 'user_likes_participant', 'user', ['user_id'], ['id'], ondelete='CASCADE')
-    op.create_foreign_key(None, 'user_likes_participant', 'participant', ['participant_id'], ['id'], ondelete='CASCADE')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'user_likes_participant', type_='foreignkey')
-    op.drop_constraint(None, 'user_likes_participant', type_='foreignkey')
-    op.create_foreign_key('user_likes_participant_ibfk_1', 'user_likes_participant', 'participant', ['participant_id'], ['id'])
-    op.create_foreign_key('user_likes_participant_ibfk_2', 'user_likes_participant', 'user', ['user_id'], ['id'])
-    op.add_column('participant', sa.Column('user_id', mysql.INTEGER(), autoincrement=False, nullable=False))
-    # ### end Alembic commands ###
