--- watchapedia/database/alembic/versions/2025_01_30_0035-5b29b6225b73_typo_fix.py
@@ -1,32 +1,0 @@
-"""typo fix
-
-Revision ID: 5b29b6225b73
-Revises: 162001d72482
-Create Date: 2025-01-30 00:35:11.630395
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = '5b29b6225b73'
-down_revision: Union[str, None] = '162001d72482'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('participant', sa.Column('likes_count', sa.Integer(), nullable=False))
-    op.drop_column('participant', 'like_count')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('participant', sa.Column('like_count', mysql.INTEGER(), autoincrement=False, nullable=False))
-    op.drop_column('participant', 'likes_count')
-    # ### end Alembic commands ###
