+++ watchapedia/database/alembic/versions/2025_01_31_1710-3a1328561655_participant_like.py
@@ -1,8 +1,8 @@
-"""participant like_count
+"""participant like
 
-Revision ID: 22af0add3b7c
+Revision ID: 3a1328561655
 Revises: b2fc27df03b7
-Create Date: 2025-01-29 23:57:46.875723
+Create Date: 2025-01-31 17:10:21.155322
 
 """
 from typing import Sequence, Union
@@ -12,7 +12,7 @@ import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
-revision: str = '22af0add3b7c'
+revision: str = '3a1328561655'
 down_revision: Union[str, None] = 'b2fc27df03b7'
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
@@ -24,20 +24,16 @@ def upgrade() -> None:
     sa.Column('id', sa.Integer(), nullable=False),
     sa.Column('user_id', sa.Integer(), nullable=False),
     sa.Column('participant_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['participant_id'], ['participant.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
+    sa.ForeignKeyConstraint(['participant_id'], ['participant.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ondelete='CASCADE'),
     sa.PrimaryKeyConstraint('id')
     )
-    op.add_column('participant', sa.Column('like_count', sa.Integer(), nullable=False))
-    op.add_column('participant', sa.Column('user_id', sa.Integer(), nullable=False))
-    op.create_foreign_key(None, 'participant', 'user', ['user_id'], ['id'], ondelete='CASCADE')
+    op.add_column('participant', sa.Column('likes_count', sa.Integer(), nullable=False))
     # ### end Alembic commands ###
 
 
 def downgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'participant', type_='foreignkey')
-    op.drop_column('participant', 'user_id')
-    op.drop_column('participant', 'like_count')
+    op.drop_column('participant', 'likes_count')
     op.drop_table('user_likes_participant')
     # ### end Alembic commands ###
