+++ watchapedia/database/alembic/versions/2025_01_31_1732-56b0a26e0de1_like_count_default_0.py
@@ -1,0 +1,30 @@
+"""like_count default 0
+
+Revision ID: 56b0a26e0de1
+Revises: 3a1328561655
+Create Date: 2025-01-31 17:32:31.867851
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '56b0a26e0de1'
+down_revision: Union[str, None] = '3a1328561655'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
